% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbeGenFns.R
\name{probeGen}
\alias{probeGen}
\title{Function to generate probes for PatH-Cap tool}
\usage{
probeGen(fasta_file, gff_file, probes_file, add_adapt = FALSE,
  pre_str = "", suf_str = "")
}
\arguments{
\item{fasta_file}{Fasta file for the bacterial genome sequence}

\item{gff_file}{Gff file for the bacterial genome annotation}

\item{probes_file}{Path to text file where probes would be written}

\item{pre_str}{[Optional] Sequence of prefix_adapter, required when add_apt is set to TRUE}

\item{suf_str}{[Optional] Sequence of suffix_adapter, required when add_apt is set to TRUE}

\item{add_apt}{[Optional] Whether adapters would be added or not}
}
\description{
This is the main function to generate probes for hybrid-capture.
}
\examples{
probeGen()
}
